Feat: separate place card component
